move random functions to a utils file
Those are cluttering up the main canned file so moving them here for
now, even though this is probably not the end goal to have...
