+++ lib/utils.js
@@ -1,0 +1,24 @@
+"use strict";
+
+var utils = module.exports = {}
+
+utils.escapeRegexSpecialChars = function (text) {
+  return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&")
+}
+
+utils.extend = function (target) {
+  var sources = [].slice.call(arguments, 1);
+  sources.forEach(function (source) {
+    for (var prop in source) {
+      if(source.hasOwnProperty(prop)){
+        target[prop] = source[prop]
+      }
+    }
+  })
+  return target
+}
+
+utils.removeJSLikeComments = function (text) {
+  return text.replace(/\/\*.+?\*\/|\/\/\s.*(?=[\n\r])/g, '')
+}
+
